requirements.txt
django
django-cleanup
pillow
requests
